refactor(rule): extract logic for mock data
