+++ src/hooks/Rule/rule/useDebugRule.ts
@@ -1,16 +1,21 @@
 import { addTrace, deleteTrace, getTraceLog } from '@/api/diagnose'
 import { applyRuleTest } from '@/api/ruleengine'
+import { getKeywordsFromSQL } from '@/common/tools'
+import useI18nTl from '@/hooks/useI18nTl'
 import useSyncPolling from '@/hooks/useSyncPolling'
 import { TraceRecord } from '@/types/diagnose'
-import { LogTraceFormatter, TraceEncodeType } from '@/types/enum'
-import { BasicRule, RuleItem } from '@/types/rule'
+import { LogTraceFormatter, RuleInputType, TraceEncodeType } from '@/types/enum'
+import { BasicRule, BridgeItem, RuleEvent, RuleItem } from '@/types/rule'
+import { ElMessageBox } from 'element-plus'
 import { cloneDeep, debounce, isArray, isEqual, isFunction, mergeWith, startCase } from 'lodash'
 import moment from 'moment'
-import type { Ref } from 'vue'
-import { computed, onMounted, onUnmounted, ref, watch } from 'vue'
+import type { ComputedRef, Ref } from 'vue'
+import { computed, onUnmounted, ref, watch } from 'vue'
 import { useStore } from 'vuex'
 import type { FormattedLog, LogItem } from './useFormatDebugLog'
 import useFormatDebugLog from './useFormatDebugLog'
+import { useRuleUtils } from './useRule'
+import useRuleEvents from './useRuleEvents'
 
 const BYTE_PER_PAGE = Math.pow(2, 30)
 
@@ -227,5 +232,175 @@ export const useStatusController = (rule?: Ref<BasicRule | RuleItem>) => {
     testTarget,
     isRuleSaveButtonDisabled,
     updateSavedRule,
+  }
+}
+
+interface TestParams {
+  context: Record<string, string>
+}
+
+export const useMockData = (
+  props: Readonly<{
+    ingressBridgeList: BridgeItem[]
+    ruleData: Record<string, any>
+  }>,
+): {
+  ruleSql: ComputedRef<string>
+  dataType: Ref<string>
+  testParams: Ref<TestParams>
+  isDataTypeNoMatchSQL: Ref<boolean>
+  eventList: Ref<Array<RuleEvent>>
+  resetContext: () => void
+  getMockContext: () => Record<string, any>
+  setDataTypeNContext: () => void
+} => {
+  const { tl, t } = useI18nTl('RuleEngine')
+
+  const eventList = ref<Array<RuleEvent>>([])
+
+  const ruleSql = computed(() => props.ruleData?.sql || '')
+  const testParams: Ref<TestParams> = ref({ context: {} })
+
+  const dataType: Ref<string> = ref('')
+  const isDataTypeNoMatchSQL = ref(false)
+
+  const { getEventList } = useRuleEvents()
+  const loadRuleEvents = async () => {
+    try {
+      eventList.value = await getEventList()
+    } catch (error) {
+      console.error(error)
+    }
+  }
+
+  const { TOPIC_EVENT, findInputTypeNTarget, getTestColumns, transFromStrToFromArr } =
+    useRuleUtils()
+
+  const findSourceTypeAndTarget = (fromTarget: string) =>
+    findInputTypeNTarget(fromTarget, eventList.value, props.ingressBridgeList)
+
+  const resetContext = () => {
+    ElMessageBox.confirm(tl('confirmReset'), {
+      confirmButtonText: t('Base.confirm'),
+      cancelButtonText: t('Base.cancel'),
+      type: 'warning',
+    })
+      .then(() => {
+        handleDataSourceChanged({ value: dataType.value })
+      })
+      .catch(() => {
+        // ignore
+      })
+  }
+
+  const setContext = (obj: Record<string, string>) => {
+    testParams.value.context = obj
+  }
+
+  const compareTargetNFromStr = (
+    targetType: RuleInputType,
+    target: BridgeItem | RuleEvent | string,
+    fromStr: string,
+  ): boolean => {
+    const inputs = transFromStrToFromArr(fromStr)
+    const typeNeedToCompares = inputs.map((input) => {
+      const { type: typeInSQL } = findSourceTypeAndTarget(input)
+      const typeNeedToCompare = typeInSQL === RuleInputType.Topic ? TOPIC_EVENT : input
+      return typeNeedToCompare
+    })
+    // when comparing, if the type is topic, compare the TOPIC_EVENT;
+    // if type is event, compare target.event
+    // if type is bridge, compare bridge.id
+    let targetStrToCompare = ''
+    switch (targetType) {
+      case RuleInputType.Topic:
+        targetStrToCompare = target as string
+        break
+      case RuleInputType.Event:
+        targetStrToCompare = (target as RuleEvent).event
+        break
+      case RuleInputType.Bridge:
+        targetStrToCompare = (target as BridgeItem).idForRuleFrom
+        break
+    }
+    return typeNeedToCompares.includes(targetStrToCompare)
+  }
+
+  const handleDataSourceChanged = ({ value }: { value: string }) => {
+    // The data type switch will not change the SQL, but if the data type does not match the SQL,
+    // a warning will be issued indicating that the data type does not match the SQL
+    const { type, target } = findSourceTypeAndTarget(value)
+    const { fromStr } = getKeywordsFromSQL(ruleSql.value)
+    isDataTypeNoMatchSQL.value = !compareTargetNFromStr(type, target, fromStr)
+    const { context } = getTestColumns(type, value, eventList.value || [])
+    setContext(context)
+  }
+
+  const handleSQLChanged = (sql: string) => {
+    const { type, target } = findSourceTypeAndTarget(dataType.value)
+    const { fromStr } = getKeywordsFromSQL(sql)
+    isDataTypeNoMatchSQL.value = !compareTargetNFromStr(type, target, fromStr)
+  }
+
+  const getEventTypeInContext = () => {
+    const { type, target } = findSourceTypeAndTarget(dataType.value)
+    if (type === RuleInputType.Event) {
+      return dataType.value.match(/(\$events\/)([\w]+)/)?.[2]
+    }
+    if (type === RuleInputType.Bridge) {
+      return `$bridges/${(target as BridgeItem).type}:*`
+    }
+    return TOPIC_EVENT.match(/(\$events\/)([\w]+)/)?.[2]
+  }
+
+  const getMockContext = () => {
+    return {
+      ...testParams.value.context,
+      event_type: getEventTypeInContext(),
+    }
+  }
+
+  const setDataType = (type: RuleInputType, firstInput: string) => {
+    if (type === RuleInputType.Topic) {
+      dataType.value = TOPIC_EVENT
+    } else {
+      dataType.value = firstInput
+    }
+  }
+  const setDataTypeNContext = () => {
+    const { fromStr } = getKeywordsFromSQL(ruleSql.value)
+    const [firstInput = ''] = transFromStrToFromArr(fromStr)
+    const { type: inputType } = findSourceTypeAndTarget(firstInput)
+    const { context } = getTestColumns(inputType, firstInput, eventList.value)
+    setDataType(inputType, firstInput)
+    testParams.value = { context }
+  }
+
+  watch(
+    () => eventList.value,
+    () => {
+      handleDataSourceChanged({ value: dataType.value })
+    },
+  )
+
+  watch(ruleSql, (val) => {
+    handleSQLChanged(val)
+  })
+
+  watch(dataType, (val) => {
+    handleDataSourceChanged({ value: val })
+  })
+
+  loadRuleEvents()
+
+  return {
+    ruleSql,
+    dataType,
+    testParams,
+    isDataTypeNoMatchSQL,
+    eventList,
+    resetContext,
+    getMockContext,
+    setDataTypeNContext,
   }
 }
